feat: remove context
